+++ docs/src/examples/Dialog/DialogWithDescriptionComponent.vue
@@ -11,9 +11,7 @@ export default {
   computed: {
     dialogConfig () {
       return {
-        card: {
-          description: DescriptionComponent
-        }
+        description: DescriptionComponent
       }
     }
   }